+++ packages/euclid/src/msgs/factory/msg.rs
@@ -3,7 +3,7 @@ use crate::{
     chain::ChainUid,
     cross_chain_user::CrossChainUser,
     fee::{DenomFees, PartnerFee},
-    liquidity::{AddLiquidityRequest, RemoveLiquidityRequest},
+    liquidity::{AddLiquidityRequest, RemoveLiquidityRequest, SingleSidedLiquidityRequest},
     msgs::{cross_chain_config::CrossChainConfig, hook::EuclidReceive, vlp::base::PoolConfig},
     recipient::Recipient,
     swap::{NextSwapPair, SwapRequest},
@@ -189,6 +189,11 @@ pub enum QueryMsg {
         user: Addr,
         pagination: Pagination<Uint256>,
     },
+    #[returns(GetPendingSingleSidedLiquidityResponse)]
+    PendingSingleSidedLiquidity {
+        user: Addr,
+        pagination: Pagination<Uint256>,
+    },
 
     #[returns(GetEscrowResponse)]
     GetEscrow { token_id: String },
@@ -285,6 +290,11 @@ pub struct GetPendingLiquidityResponse {
 #[cw_serde]
 pub struct GetPendingRemoveLiquidityResponse {
     pub pending_remove_liquidity: Vec<RemoveLiquidityRequest>,
+}
+
+#[cw_serde]
+pub struct GetPendingSingleSidedLiquidityResponse {
+    pub pending_single_sided_liquidity: Vec<SingleSidedLiquidityRequest>,
 }
 
 #[cw_serde]